Conflicts:
	tests/Sabre/DAV/Mock/Collection.php
	tests/Sabre/DAV/ServerSimpleTest.php

+++ lib/Sabre/DAV/Server.php
@@ -684,6 +684,9 @@ class Server {
      */
     protected function httpDelete($uri) {
 
+        // Checking If-None-Match and related headers.
+        if (!$this->checkPreconditions()) return;
+
         if (!$this->broadcastEvent('beforeUnbind',array($uri))) return;
         $this->tree->delete($uri);
         $this->broadcastEvent('afterUnbind',array($uri));
